+++ include/WCSimWCAddDarkNoise.hh
@@ -23,8 +23,8 @@ public:
   
 public:
   void AddDarkNoise();
-  void AddDarkNoiseBeforeDigi(WCSimWCDigitsCollection* WCHCPMT, float num1 ,float num2);
-  void FindDarkNoiseRanges(WCSimWCDigitsCollection* WCHCPMT, float width);
+  void AddDarkNoiseBeforeDigi(WCSimWCDigitsCollection* WCHCPMT, double num1 ,double num2);
+  void FindDarkNoiseRanges(WCSimWCDigitsCollection* WCHCPMT, double width);
   //As it inherits from G4VDigitizerModule it needs a digitize class.  Not used
   void Digitize() { }
   void SetDarkRate(double idarkrate){ PMTDarkRate = idarkrate; }
@@ -56,8 +56,8 @@ private:
   WCSimDetectorConstruction* myDetector;
   G4String detectorElement;
   
-  std::vector<std::pair<float, float> > ranges;
-  std::vector<std::pair<float, float> > result;
+  std::vector<std::pair<double, double> > ranges;
+  std::vector<std::pair<double, double> > result;
   
 };
